app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.4.2
python-multipart==0.0.6
mysql-connector-python==8.1.0
joblib==1.3.2
